Add "Armor Plating" Redeem
Adds a new redeem, "Armor Plating", which adds a defensive boost
behaving similar to the Speed Boost ability.

When selecting this redeem, the user is prompted to choose between
Defense and Sp. Def. This distinction is stored in the ROM as a
flag, `0x4AE`. The redeem itself is saved in the same
`VAR_ROGUEMON_REDEEMS` var as Revive and Goody Jar.