Add click to view trainer info
Adds an overlay to the game display which, when clicked, shows the
the info for the trainer currently under the cursor. The algorithm
for doing this is as follows:
 * Get the on-screen coordinates of the mouse click event
 * Read the player's in-game coordinates from the game memory
    * Assumes the player is always in the middle of the screen: 7 tiles
      from the left and 4.5 tiles from the top
 * Use the player's position to calculate the tile under the cursor
 * Check a list of all trainers in the current current area to see
   which, if any, trainer is on the clicked tile

The list of trainers in each area is stored in `roguemon/trainers.lua`.
This file was generated using a script which does the following:
 * Extract a list of trainers from each area's `map.json`
 * Read the script assigned to each trainer
 * Parse the area's event script to find the trainer ID
 * Read the `trainers.h` header to find the numeric value of that ID

`trainers.lua` stores a mapping of these trainer IDs to their
coordinates; it also includes a helper function, `CheckCoord()`, to find
which trainer ID matches a given coordinate.